Домашка

+++ t3.java
@@ -5,13 +5,13 @@ public class t3 {
         Scanner sc = new Scanner(System.in);
 
         System.out.print("Enter first number: ");
-        float a = sc.nextFloat();
+        double a = sc.nextDouble();
 
         System.out.print("Enter operation (*, /, +, -): ");
         char operator = sc.next().charAt(0);
 
         System.out.print("Enter second number: ");
-        float b = sc.nextFloat();
+        double b = sc.nextDouble();
 
         switch (operator) {
             case '*' -> System.out.printf("Result: %f", a * b);
